separate CellMetadata Class to its own file
moved validation directory into ingest
